requirements.txt
aiobotocore==2.20.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
boto3==1.37.0
botocore==1.37.0
frozenlist==1.5.0
fsspec==2025.2.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
numpy==2.0.2
pandas==2.2.3
propcache==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
s3fs==2025.2.0
s3transfer==0.11.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wrapt==1.17.2
yarl==1.18.3
